# wafaa/app/Models/Artist.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Artist extends Model
{
    use HasFactory;

     protected $fillable = [
        'name',
        'genre',
        'country',
    ];

    public function albums()
{
    return $this->hasMany(Album::class);
}

}